requirements.txt
Flask==3.1.1
cloudinary==1.44.1
requests==2.32.4
Pillow
gunicorn